# src/GraffLibAPI/GraffLibAPI/GraffLibAPI/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a GraffLib server that exposes various endpoints for client app.",
    "version": "1.0.0",
    "title": "GraffLib REST API"
  },
  "host": "localhost:8000",
  "basePath": "/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user.",
        "description": "",
        "operationId": "postUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the system.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation.",
            "schema": {
              "$ref": "#/definitions/CreateUserResponse"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "409": {
            "description": "User already exists."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/users/password-recovery": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Send a password recovery email.",
        "description": "",
        "operationId": "sendPasswordRecoveryEmail",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SendUserPasswordRecoveryRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted operation."
          },
          "400": {
            "description": "Bad request."
          },
          "404": {
            "description": "User was not found."
          },
          "405": {
            "description": "Validation exception."
          },
          "409": {
            "description": "User does not support this operation."
          },
          "500": {
            "description": "Internal server error."
          }

        }
      }
    },
    "/users/password": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Update user's password after sending recovery email.",
        "description": "",
        "operationId": "updateUserPasswordAfterRecoveryEmail",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserPasswordRecoveryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User will be redirected to html template with success message."
          },
          "400": {
            "description": "User will be redirected to html template with failure message."
          },
          "404": {
            "description": "User will be redirected to html template with failure message."
          },
          "405": {
            "description": "User will be redirected to html template with failure message."
          },
          "409": {
            "description": "User will be redirected to html template with failure message."
          },
          "500": {
            "description": "User will be redirected to html template with failure message."
          }
        }
      },

      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user's password when user is authenticated.",
        "description": "",
        "operationId": "updateUserPasswordUserAuthenticated",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserPasswordUserAuthenticatedRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "User is not authenticated."
          },
          "403": {
            "description": "User is not authorized."
          },
          "409": {
            "description": "There are problems with old/new passwords."
          },
          "422": {
            "description": "User can't update other user passwords, only his/her owns."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/admins/users": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "Get all users.",
        "description": "",
        "operationId": "getAllUsersByAdmin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "description": "User is not authenticated."
          },
          "403": {
            "description": "User is not authorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "Get a user by UserId.",
        "description": "",
        "operationId": "getUserByIdByAdmin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of User to return.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "User is not authenticated."
          },
          "403": {
            "description": "User is not authorized."
          },
          "404": {
            "description": "User was not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "tags": [
          "Admins"
        ],
        "summary": "Delete a specific user.",
        "description": "",
        "operationId": "deleteUserByAdmin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of User to delete.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "401": {
            "description": "User is not authenticated."
          },
          "403": {
            "description": "User is not authorized."
          },
          "404": {
            "description": "User was not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "description": "user privillege status",
          "enum": [
            "user",
            "admin"
          ]
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "properties": {
        "user_name": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "description": "user privillege status",
          "enum": [
            "user",
            "admin"
          ]
        }
      }
    },
    "CreateUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "SendUserPasswordRecoveryRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "UpdateUserPasswordRecoveryRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      }
    },
    "UpdateUserPasswordUserAuthenticatedRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "integer",
          "format": "int64"
        },
        "old_password": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        }
      }
    },
...
  },
...
}